--- TODO.md
@@ -1,3 +1,0 @@
-1. Autentisering
-2. Loading states
-3. Feilhåndtering
